# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GreenHerb API
  description: |
    API REST para a gestão de planos de cultivo da estufa GreenHerb.
    Projeto académico — Sprint 2.
  version: 0.1.0
  contact:
    name: Equipa GreenHerb
servers:
  - url: http://localhost:5000
    description: Servidor de desenvolvimento local
...
paths:
...
  /api/auth/register:
    post:
      tags: [Autenticação]
      summary: Registar novo utilizador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Utilizador criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/auth/login:
    post:
      tags: [Autenticação]
      summary: Iniciar sessão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Sessão iniciada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/ervas:
    get:
      tags: [Ervas]
      summary: Listar ervas (público)
      responses:
        '200':
          description: Lista de ervas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ErvaAromatica'
    post:
      tags: [Ervas]
      summary: Criar erva (Administrador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErvaAromatica'
      responses:
        '201':
          description: Erva criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErvaAromatica'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/ervas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Ervas]
      summary: Editar erva (Administrador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ErvaAromatica'
      responses:
        '200':
          description: Erva atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErvaAromatica'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Ervas]
      summary: Eliminar erva (Administrador)
      security:
        - bearerAuth: []
      responses:
        '204': { description: Erva eliminada }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/ervas/importar:
    post:
      tags: [Ervas]
      summary: Importar ervas via CSV ou Excel (Administrador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [ficheiro]
              properties:
                ficheiro:
                  type: string
                  format: binary
                  description: Ficheiro .csv ou .xlsx
      responses:
        '200':
          description: Resumo da importação
          content:
            application/json:
              schema:
                type: object
                properties:
                  inseridas: { type: integer }
                  ignoradas: { type: integer }
                  erros:
                    type: array
                    items:
                      type: object
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /api/alertas:
    get:
      tags: [Alertas]
      summary: Listar alertas
      security:
        - bearerAuth: []
      parameters:
        - name: estado
          in: query
          schema:
            type: string
            enum: [ativo, resolvido, ignorado]
        - name: loteId
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Alerta'
  /api/alertas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Alertas]
      summary: Detalhe do alerta
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Alerta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alerta'
  /api/alertas/{id}/resolver:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Alertas]
      summary: Resolver alerta
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Alerta resolvido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alerta'
  /api/alertas/{id}/ignorar:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Alertas]
      summary: Ignorar alerta com justificação
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [justificacao]
              properties:
                justificacao:
                  type: string
                  minLength: 10
      responses:
        '200':
          description: Alerta ignorado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alerta'
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Pedido inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Autenticação requerida ou token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sem permissões para a operação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        erro:
          type: string
          example: Mensagem descritiva do erro.
    Utilizador:
      type: object
      properties:
        _id:
          type: string
          example: 65a1b2c3d4e5f60718293a4b
        nome:
          type: string
          example: Ana Administradora
        email:
          type: string
          format: email
          example: admin@greenherb.local
        perfil:
          type: string
          enum: [Técnico, Responsável, Administrador]
          example: Administrador
        ativo:
          type: boolean
          example: true
        dataCriacao:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [nome, email, password]
      properties:
        nome:
          type: string
          example: Tânia Técnica
        email:
          type: string
          format: email
          example: tecnico@greenherb.local
        password:
          type: string
          format: password
          minLength: 6
          example: tecnico123
        perfil:
          type: string
          enum: [Técnico, Responsável, Administrador]
          default: Técnico
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@greenherb.local
        password:
          type: string
          format: password
          example: admin123
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT a usar no header Authorization.
        utilizador:
          $ref: '#/components/schemas/Utilizador'
...
    ErvaAromatica:
      type: object
      required: [nome]
      properties:
        _id: { type: string }
        nome: { type: string, example: Manjericão }
        descricao: { type: string }
        tempMin: { type: number }
        tempMax: { type: number }
        humidadeMin: { type: number, minimum: 0, maximum: 100 }
        humidadeMax: { type: number, minimum: 0, maximum: 100 }
        luminosidadeMin: { type: number, minimum: 0 }
        luminosidadeMax: { type: number, minimum: 0 }
        dataCriacao: { type: string, format: date-time }
...
    Alerta:
      type: object
      required: [mensagem]
      properties:
        _id: { type: string }
        loteId: { type: string }
        medicaoId: { type: string }
        classificacao:
          type: string
          enum: [Informativo, Aviso, Crítico]
        mensagem: { type: string }
        estado:
          type: string
          enum: [ativo, resolvido, ignorado]
        justificacaoIgnorar: { type: string }
        resolvidoPor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Utilizador'
        dataRegisto: { type: string, format: date-time }
        dataResolucao: { type: string, format: date-time }
...